# Replaced classes with functions