--- a/woltka/align.py
+++ b/woltka/align.py
@@ -26,47 +26,86 @@ A parser function returns a tuple of:
 from collections import deque
 
 
-def plain_mapper(fh, fmt=None, n=None):
-    n = n or 1000000  # default chunk size: 1 million
-    i = 0        # current line number
-    j = n        # target line number at end of current chunk
+def plain_mapper(fh, fmt=None, n=1000000):
+    """Read an alignment file in chunks and yield query-to-subject(s) maps.
 
-    # determine file format based on first line
-    if fmt is None:
-        try:
-            line = next(fh)
-        except StopIteration:
-            return
-        fmt = infer_align_format(line)
-        fh.seek(0)
+    Parameters
+    ----------
+    fh : file handle
+        Alignment file to parse.
+    fmt : str, optional
+        Alignment file format.
+    n : int, optional
+        Number of lines per chunk.
+
+    Yields
+    ------
+    deque of str
+        Query queue.
+    deque of set of str
+        Subject(s) queue.
+
+    Notes
+    -----
+    The design of this function aims to couple with the extremely large size of
+    typical alignment files. It reads the entire file sequentially, pauses and
+    processes current cache for every _n_ lines, yields and clears cache, then
+    proceeds.
+    """
+    # determine alignment file format
+    fmt = fmt or infer_align_format(fh)
 
     # assign parser for given format
     parser = assign_parser(fmt)
 
+    # query queue and subject(s) queue
     qryque, subque = deque(), deque()
+
+    # pre-load method references
     qry_append, sub_append = qryque.append, subque.append
-    qry_clear, sub_clear = qryque.clear, subque.clear
 
     # parse alignment file
     this = None  # current query Id
+    target = n   # target line number at end of current chunk
     for i, line in enumerate(fh):
 
+        # parse current alignment line
         try:
             query, subject = parser(line)[:2]
         except TypeError:
             continue
 
+        # add subject to subject set of the same query Id
         if query == this:
             subque[-1].add(subject)
+
+        # when query Id changes,
         else:
-            if i > j:
+
+            # line number has reached target
+            if i > target:
+
+                # flush current queries and subject sets
+                # this happens only when: 1) query Id changes, and 2) line
+                # number has reached target, so that subjects of the same query
+                # won't be separated in multiple flushes
                 yield qryque, subque
-                qry_clear()
-                sub_clear()
-                j = i + n
+
+                # re-initiate queues and method references
+                qryque, subque = deque(), deque()
+                qry_append, sub_append = qryque.append, subque.append
+
+                # next target line number
+                target = i + n
+
+            # create new query and subject set pair
             qry_append(query)
             sub_append({subject})
+
+            # update current query Id
             this = query
+
+    # final flush
     yield qryque, subque
 
 
@@ -228,21 +267,23 @@ class Plain(object):
             self.__init__()
 
 
-def infer_align_format(line):
+def infer_align_format(fh):
     """Guess the format of an alignment file based on first line.
 
     Parameters
     ----------
-    line : str
-        First line of alignment file.
+    fh : file handle
+        Input alignment file.
 
     Returns
     -------
-    str or None
+    str
         Alignment file format (map, b6o or sam).
 
     Raises
     ------
+    ValueError
+        Alignment file cannot be read.
     ValueError
         Alignment file format cannot be determined.
 
@@ -251,17 +292,39 @@ def infer_align_format(line):
     parse_b6o_line
     parse_sam_line
     """
+    # read first line of file
+    try:
+        line = next(fh)
+
+    # file is empty or not readable
+    except StopIteration:
+        raise ValueError('Cannot read alignment file.')
+
+    # move pointer back to beginning of file
+    fh.seek(0)
+
+    # SAM file header
     if line.split()[0] in ('@HD', '@PG'):
         return 'sam'
+
+    # tab-delimited row
     row = line.rstrip().split('\t')
+
+    # simple query-to-subject map
     if len(row) == 2:
         return 'map'
+
+    # BLAST standard tabular format
     if len(row) >= 12:
         if all(row[i].isdigit() for i in range(3, 10)):
             return 'b6o'
+
+    # SAM format
     if len(row) >= 11:
         if all(row[i].isdigit() for i in (1, 3, 4)):
             return 'sam'
+
+    # cannot determine
     raise ValueError('Cannot determine alignment file format.')