--- a/woltka/workflow.py
+++ b/woltka/workflow.py
@@ -19,20 +19,21 @@ output (via `click`) and file input/output, except for raising errors.
 from os import makedirs
 from os.path import join, basename, isfile, isdir
 from collections import deque
+from functools import partial
 import click
 
 from .util import update_dict, allkeys, sum_dict, intize
 from .file import (
     openzip, path2stem, read_ids, id2file_from_dir, id2file_from_map, read_map,
     write_readmap, write_table)
-from .align import Plain, parse_align_file
+from .align import Plain, parse_align_file, plain_mapper
 from .classify import (
     assign_none, assign_free, assign_rank, count, count_strata, strip_index,
     demultiplex)
 from .tree import (
     read_names, read_nodes, read_lineage, read_newick, read_rank_table,
     fill_root)
-from .ordinal import Ordinal, read_gene_coords, whether_prefix
+from .ordinal import Ordinal, read_gene_coords, whether_prefix, ordinal_mapper
 from .biom import profile_to_biom, write_biom
 
 
@@ -221,7 +222,8 @@ def classify(mapper:  object,
         with openzip(fp) as fh:
 
             # parse alignment file by chunk
-            for qryque, subque in parse_align_file(fh, mapper, fmt, lines):
+            # for qryque, subque in parse_align_file(fh, mapper, fmt, lines):
+            for qryque, subque in mapper(fh, fmt=fmt, n=lines):
 
                 # show progress
                 click.echo('.', nl=False)
@@ -418,10 +420,14 @@ def build_mapper(coords_fp: str = None,
             coords = read_gene_coords(fh, sort=True)
         click.echo(' Done.')
         click.echo(f'Total number of host sequences: {len(coords)}.')
-        return Ordinal(coords, whether_prefix(coords),
-                       overlap and overlap / 100)
+        # return Ordinal(coords, whether_prefix(coords),
+        #                overlap and overlap / 100)
+        return partial(ordinal_mapper, coords=coords,
+                       prefix=whether_prefix(coords),
+                       th=overlap and overlap / 100)
     else:
-        return Plain()
+        # return Plain()
+        return plain_mapper
 
 
 def prepare_ranks(ranks:      str = None,