--- a/woltka/workflow.py
+++ b/woltka/workflow.py
@@ -26,14 +26,14 @@ from .util import update_dict, allkeys, sum_dict, intize
 from .file import (
     openzip, path2stem, read_ids, id2file_from_dir, id2file_from_map, read_map,
     write_readmap, write_table)
-from .align import Plain, parse_align_file, plain_mapper
+from .align import plain_mapper
 from .classify import (
     assign_none, assign_free, assign_rank, count, count_strata, strip_index,
     demultiplex)
 from .tree import (
     read_names, read_nodes, read_lineage, read_newick, read_rank_table,
     fill_root)
-from .ordinal import Ordinal, read_gene_coords, whether_prefix, ordinal_mapper
+from .ordinal import ordinal_mapper, read_gene_coords, whether_prefix
 from .biom import profile_to_biom, write_biom
 
 
@@ -391,8 +391,8 @@ def parse_strata(fp:       str = None,
 
 
 def build_mapper(coords_fp: str = None,
-                 overlap:   int = None) -> object:
-    """Build mapping module (Plain or Ordinal).
+                 overlap:   int = None) -> callable:
+    """Build mapper function (plain or ordinal).
 
     Parameters
     ----------
@@ -403,8 +403,8 @@ def build_mapper(coords_fp: str = None,
 
     Returns
     -------
-    object
-        Mapping module.
+    callable
+        Mapper function.
 
     Notes
     -----
@@ -420,13 +420,10 @@ def build_mapper(coords_fp: str = None,
             coords = read_gene_coords(fh, sort=True)
         click.echo(' Done.')
         click.echo(f'Total number of host sequences: {len(coords)}.')
-        # return Ordinal(coords, whether_prefix(coords),
-        #                overlap and overlap / 100)
         return partial(ordinal_mapper, coords=coords,
                        prefix=whether_prefix(coords),
                        th=overlap and overlap / 100)
     else:
-        # return Plain()
         return plain_mapper